[Refactor] : ApplicationNotificationCenter 생성
Related to: #35
